refactor(v2): extract object store package
